move TokenType enum to separate file
