+++ src/Domain/TokenType.ts
@@ -1,0 +1,5 @@
+export enum TokenType {
+    REMOVABLE = "removable",
+    SELECTABLE = "selectable",
+    NONEXISTENT = "nonexistent",
+}
